Vim-plugin recipe better handles std/opt packages
- When a plugin has already been loaded into vim's standard or optional
paths and you decide to change it to be loaded into the alternate path,
we were leaving the artifact behind and simply fetching the remote into
the new destination. This change enables a check for a local variant
before fetching so that it is lifted from the old location to the new
one before refreshing it from the origin.